[feat] move tests from /test to /actions/order
