+++ packages/core/src/actions/order/cancelOffer.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 
 import { accounts, config, mintERC20, mintERC721 } from "@ark-project/test";
 
-import { cancelOrder, createOffer, getOrderStatus } from "../src/index.js";
+import { cancelOrder, createOffer, getOrderStatus } from "../../index.js";
 
 describe("cancelOffer", () => {
   it("default", async () => {
