+++ packages/core/src/actions/order/createAuction.test.ts
@@ -2,8 +2,8 @@ import { describe, expect, it } from "vitest";
 
 import { accounts, config, mintERC721 } from "@ark-project/test";
 
-import { createAuction, getOrderType } from "../src/index.js";
-import { FEE_TOKEN, getTypeFromCairoCustomEnum } from "./utils/index.js";
+import { createAuction, getOrderType } from "../../index.js";
+import { FEE_TOKEN, getTypeFromCairoCustomEnum } from "../../../tests/utils/index.js";
 
 describe("createAuction", () => {
   it("default", async () => {
