+++ packages/core/src/actions/order/fulfillAuction.test.ts
@@ -8,9 +8,9 @@ import {
   mintERC721
 } from "@ark-project/test";
 
-import { fulfillAuction } from "../src/actions/order/fulfillAuction.js";
-import { createAuction, createOffer } from "../src/actions/order/index.js";
-import { getOrderStatus } from "../src/actions/read/index.js";
+import { fulfillAuction } from "../../index.js";
+import { createAuction, createOffer } from "../../index.js";
+import { getOrderStatus } from "../read/index.js";
 
 describe("fulfillAuction", () => {
   it("default", async () => {
